Toast data interfaces

+++ tests/_helpers.py
@@ -15,9 +15,13 @@ try:
     import toast
     import sotodlib.toast as sotoast
     import sotodlib.toast.ops as so_ops
+    import toast.schedule_sim_ground
+    from toast.observation import default_values as defaults
     toast_available = True
 except ImportError as e:
     toast_available = False
+
+from sotodlib.sim_hardware import sim_nominal
 
 
 def mpi_world():
@@ -48,20 +52,20 @@ def mpi_multi():
         return False
 
 
-def create_outdir(subdir=None, comm=None):
+def create_outdir(subdir=None, mpicomm=None):
     """Create the top level output directory and per-test subdir.
 
     Args:
         subdir (str): the sub directory for this test.
-        comm (MPI.Comm):  Optional communicator.
+        mpicomm (MPI.Comm):  Optional communicator.
 
     Returns:
         str: full path to the test subdir if specified, else the top dir.
 
     """
     rank = 0
-    if comm is not None:
-        rank = comm.rank
+    if mpicomm is not None:
+        rank = mpicomm.rank
     retdir = None
     if rank == 0:
         pwd = os.path.abspath(".")
@@ -73,8 +77,8 @@ def create_outdir(subdir=None, comm=None):
             os.mkdir(testdir)
         if not os.path.isdir(retdir):
             os.mkdir(retdir)
-    if comm is not None:
-        retdir = comm.bcast(retdir, root=0)
+    if mpicomm is not None:
+        retdir = mpicomm.bcast(retdir, root=0)
     return retdir
 
 
@@ -106,55 +110,45 @@ def close_data_and_comm(data):
     del cm
 
 
-# FIXME:  PR #183 has additional helper functions for instrument classes.
-# Just use those once that is merged and delete this.
-def toast_site():
-    if not toast_available:
-        print("toast unavailable- cannot create site")
-        return
-    return toast.instrument.GroundSite(
-        "ATACAMA",
-        -22.958 * u.degree,
-        -67.786 * u.degree,
-        5200.0 * u.meter,
-    )
-
-
-def observing_schedule(telescope, temp_dir=None):
+def observing_schedule(telescope, mpicomm=None, temp_dir=None):
     if not toast_available:
         print("toast unavailable- cannot build observing schedule")
         return
-    tdir = temp_dir
-    if tdir is None:
-        tdir = tempfile.mkdtemp()
+    schedule = None
+    if mpicomm is None or mpicomm.rank == 0:
+        tdir = temp_dir
+        if tdir is None:
+            tdir = tempfile.mkdtemp()
 
-    sch_file = os.path.join(tdir, "schedule.txt")
-    toast.schedule_sim_ground.run_scheduler(
-        opts=[
-            "--site-name",
-            telescope.site.name,
-            "--telescope",
-            telescope.name,
-            "--site-lon",
-            "{}".format(telescope.site.earthloc.lon.to_value(u.degree)),
-            "--site-lat",
-            "{}".format(telescope.site.earthloc.lat.to_value(u.degree)),
-            "--site-alt",
-            "{}".format(telescope.site.earthloc.height.to_value(u.meter)),
-            "--patch",
-            "small_patch,1,40,-40,44,-44",
-            "--start",
-            "2025-01-01 00:00:00",
-            "--stop",
-            "2025-01-01 00:10:00",
-            "--out",
-            sch_file,
-        ]
-    )
-    schedule = toast.schedule.GroundSchedule()
-    schedule.read(sch_file)
-    if temp_dir is None:
-        shutil.rmtree(tdir)
+        sch_file = os.path.join(tdir, "schedule.txt")
+        toast.schedule_sim_ground.run_scheduler(
+            opts=[
+                "--site-name",
+                telescope.site.name,
+                "--telescope",
+                telescope.name,
+                "--site-lon",
+                "{}".format(telescope.site.earthloc.lon.to_value(u.degree)),
+                "--site-lat",
+                "{}".format(telescope.site.earthloc.lat.to_value(u.degree)),
+                "--site-alt",
+                "{}".format(telescope.site.earthloc.height.to_value(u.meter)),
+                "--patch",
+                "small_patch,1,40,-40,44,-44",
+                "--start",
+                "2025-01-01 00:00:00",
+                "--stop",
+                "2025-01-01 02:00:00",
+                "--out",
+                sch_file,
+            ]
+        )
+        schedule = toast.schedule.GroundSchedule()
+        schedule.read(sch_file)
+        if temp_dir is None:
+            shutil.rmtree(tdir)
+    if mpicomm is not None:
+        schedule = mpicomm.bcast(schedule, root=0)
     return schedule
 
 
@@ -181,15 +175,6 @@ def calibration_schedule(telescope):
             az_min=-1.0 * u.degree,
             az_max=1.0 * u.degree,
             el=60.0 * u.degree,
-            sun_az_begin=180 * u.degree,
-            sun_az_end=180 * u.degree,
-            sun_el_begin=20 * u.degree,
-            sun_el_end=20 * u.degree,
-            moon_az_begin=180 * u.degree,
-            moon_az_end=180 * u.degree,
-            moon_el_begin=20 * u.degree,
-            moon_el_end=20 * u.degree,
-            moon_phase=0.0,
             scan_indx=0,
             subscan_indx=0,
         )
@@ -205,3 +190,179 @@ def calibration_schedule(telescope):
     )
 
     return schedule
+
+
+def create_comm(mpicomm):
+    """Create a toast communicator.
+
+    Use the specified MPI communicator to attempt to create 2 process groups.
+    If less than 2 processes are used, create a single process group.
+
+    Args:
+        mpicomm (MPI.Comm): the MPI communicator (or None).
+
+    Returns:
+        toast.Comm: the 2-level toast communicator.
+
+    """
+    if not toast_available:
+        raise RuntimeError("TOAST is not importable, cannot create a toast.Comm")
+    toastcomm = None
+    if mpicomm is None:
+        toastcomm = toast.Comm(world=mpicomm)
+    else:
+        worldsize = mpicomm.size
+        groupsize = 1
+        if worldsize >= 2:
+            groupsize = worldsize // 2
+        toastcomm = toast.Comm(world=mpicomm, groupsize=groupsize)
+    return toastcomm
+
+
+def simulation_test_data(
+    mpicomm,
+    telescope_name="SAT4",
+    wafer_slot="w42",
+    bands=["SAT_f030",],
+    sample_rate=10.0 * u.Hz,
+    detset_key="pixel",
+    temp_dir=None,
+    el_nod=False,
+    el_nods=[-1 * u.degree, 1 * u.degree],
+    thin_fp=4,
+    cal_schedule=False,
+):
+    """Create a data object with a simple ground sim.
+
+    Use the specified MPI communicator to attempt to create 2 process groups.  Create
+    a fake telescope and run the ground sim to make some observations for each
+    group.  This is useful for testing many operators that need some pre-existing
+    observations with boresight pointing.
+
+    Args:
+        mpicomm (MPI.Comm): the MPI communicator (or None).
+        sample_rate (Quantity): the sample rate.
+
+    Returns:
+        toast.Data: the distributed data with named observations.
+
+    """
+    if not toast_available:
+        raise RuntimeError("TOAST is not importable, cannot simulate test data")
+
+    # Create the communicator
+    toastcomm = create_comm(mpicomm)
+
+    hwp_rpm = 120.0
+    hwp_name = defaults.hwp_angle
+    if telescope_name == "LAT":
+        hwp_name = None
+        hwp_rpm = None
+
+    # Simulated telescope
+    telescope = sotoast.simulated_telescope(
+        hw=None,
+        hwfile=None,
+        telescope_name=telescope_name,
+        sample_rate=sample_rate,
+        bands=bands,
+        wafer_slots=wafer_slot,
+        thinfp=thin_fp,
+        comm=mpicomm,
+    )
+
+    data = toast.Data(toastcomm)
+
+    # Create a schedule.
+
+    if cal_schedule:
+        schedule = calibration_schedule(telescope)
+    else:
+        schedule = observing_schedule(telescope, mpicomm=toastcomm.comm_world)
+
+    sim_ground = toast.ops.SimGround(
+        name="sim_ground",
+        telescope=telescope,
+        session_split_key="tele_wf_band",
+        schedule=schedule,
+        hwp_angle=hwp_name,
+        hwp_rpm=hwp_rpm,
+        weather="atacama",
+        median_weather=True,
+        detset_key=detset_key,
+        elnod_start=el_nod,
+        elnods=el_nods,
+        scan_accel_az=3 * u.degree / u.second ** 2,
+    )
+    sim_ground.apply(data)
+    return data
+
+
+def simulation_test_multitube(
+    mpicomm,
+    telescope_name="LAT",
+    tubes=None,
+    sample_rate=1.0 * u.Hz,
+    detset_key="pixel",
+    temp_dir=None,
+    el_nod=False,
+    el_nods=[-1 * u.degree, 1 * u.degree],
+    thin_fp=30,
+):
+    """Create a data object with a simple ground sim.
+
+    This function generates data for a complete telescope at low sample rate, in order
+    to test operations that need this.
+
+    Args:
+        mpicomm (MPI.Comm): the MPI communicator (or None).
+        telescope_name (str):  the telescope to simulate.
+        sample_rate (Quantity): the sample rate.
+
+    Returns:
+        toast.Data: the distributed data with named observations.
+
+    """
+    if not toast_available:
+        raise RuntimeError("TOAST is not importable, cannot simulate test data")
+
+    # Create the communicator with a single group
+    toastcomm = toast.Comm(world=mpicomm)
+
+    # Simulated telescope
+    telescope = sotoast.simulated_telescope(
+        hw=None,
+        hwfile=None,
+        telescope_name=telescope_name,
+        tube_slots=tubes,
+        sample_rate=sample_rate,
+        thinfp=thin_fp,
+        comm=mpicomm,
+    )
+
+    data = toast.Data(toastcomm)
+
+    # Create a schedule.
+
+    schedule = observing_schedule(telescope, mpicomm=toastcomm.comm_world)
+
+    hwp_name = None
+    if telescope_name != "LAT":
+        hwp_name = defaults.hwp_angle
+
+    sim_ground = toast.ops.SimGround(
+        name="sim_ground",
+        telescope=telescope,
+        session_split_key="tele_wf_band",
+        schedule=schedule,
+        weather="atacama",
+        median_weather=True,
+        detset_key=detset_key,
+        elnod_start=el_nod,
+        elnods=el_nods,
+        scan_accel_az=3 * u.degree / u.second ** 2,
+        hwp_angle=hwp_name,
+    )
+    sim_ground.apply(data)
+
+    return data
